Move stores from /src/ui/stores to /src/stores
