+++ src/ui/index.js
@@ -7,7 +7,7 @@ import {
 import HomeState from './states/home/HomeState'
 import ImportIdentityState from './states/importIdentity/ImportIdentityState'
 import './styles/app.pcss'
-import { useIdentitiesStore } from './stores/identitiesStore'
+import { useIdentitiesStore } from '../stores/identitiesStore'
 import { useSdk } from './hooks/useSdk'
 import { useChromeStorage } from './hooks/useChromeStorage'
 import ApproveTransactionState from './states/approveTransaction/ApproveTransactionState'
