+++ src/ui/states/approveTransaction/ApproveTransactionState.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router'
 import { base64 as base64Decoder } from '@scure/base'
 import { useSdk } from '../../hooks/useSdk'
-import { useIdentitiesStore } from '../../stores/identitiesStore'
+import { useIdentitiesStore } from '../../../stores/identitiesStore'
 import hash from 'hash.js'
 import './approve.transaction.css'
 import TransactionDetails from './TransactionDetails'
